+++ flutter/digit-ui-components/storybook/lib/widgets/atoms/toast_stories.dart
@@ -58,6 +58,7 @@ List<Story> toastStories() {
           Toast.showToast(context,
               message: 'This is a toast!',
               type: ToastType.info,
+            position: ToastPosition.aboveOneButtonFooter
           );
         },
         child: const Text("Show Toast"),